move warden and config class outside of v1 level. refactor code and test
