+++ spec/vcr/Volcanic_Authenticator_Warden/_validate_token_always/success/1_2_1_1.yml
