+++ spec/vcr/Volcanic_Authenticator_Warden/validate_token_present/success/1_3_1_1.yml
